begin refactor, move widgets
