+++ rivers2stratigraphy/__init__.py
@@ -1,3 +1,2 @@
-from . import launcher
-
-launcher.import_runner()
+# from . import launcher
+# launcher.import_runner()
